Prepare osc framework for bigcount
Update the osc framework to use size_t for counts and ptrdiff_t for
displacements.

+++ ompi/mca/osc/base/osc_base_init.c
@@ -34,7 +34,7 @@ int
 ompi_osc_base_select(ompi_win_t *win,
                      void **base,
                      size_t size,
-                     int disp_unit,
+                     ptrdiff_t disp_unit,
                      ompi_communicator_t *comm,
                      int flavor,
                      int *model)
